move helper functions to spec_helper
